requirements.txt
numpy
pandas
matplotlib
torch
torchnet
torchvision
visdom
pillow
opencv-python
tqdm
h5py
fire